chunked out individual tests from test files
