+++ test_files/test_belongsto.js
@@ -1,0 +1,8 @@
+module.exports.run = function()
+{
+	successes = 0
+	failures = 0
+	console.log('Test: belongsto')
+	condition(1108730699, [102191581, 85633287, 85688877])
+	console.log('Successes:'.green, successes, 'Failures:'.red, failures)
+}
