+++ test_files/test_isparent.js
@@ -1,26 +1,3 @@
-// 85922583 includes 85688637 as region_id in wof:hierarchy
-// curl -X GET 'https://whosonfirst.mapzen.com/api/rest/?method=whosonfirst.places.getInfo&access_token=f4524f4adde9f758a02da8e3d03293ee&id=85922583&extras=wof:hierarchy'
-
-// https://whosonfirst.mapzen.com/api/rest/?method=whosonfirst.places.getInfo&access_token=f4524f4adde9f758a02da8e3d03293ee&id=85922583&extras=wof:hierarchy
-
-// {"record":{"wof:id":85922583,"wof:parent_id":"102087579","wof:name":"San Francisco","wof:placetype":"locality","wof:country":"US","wof:repo":"whosonfirst-data","wof:hierarchy":[{"continent_id":102191575,"locality_id":85922583,"country_id":85633793,"region_id":85688637,"county_id":102087579}]},"stat":"ok"}% 
-
-var successes = 0
-var failures = 0
-var get_record = require('../utils/get_record.js')
-// pulling everything together
-function isparent(recordid, parentlayer, parentid)
-{
-    var result = get_record.get_record_from_disk(recordid.toString())
-    if (result.properties["wof:hierarchy"][0][parentlayer + '_id'] === parentid) {
-    	console.log('Test passed.'.green)
-    	successes ++
-    }
-    	else {console.log('Test failed.'.red)
-    	failures ++
-    }
-}
-
 module.exports.run = function()
 {
 	successes = 0
