+++ test_runs/test_belongsto.js
@@ -19,11 +19,11 @@ function condition(recordid, belongsto_ids)
     }
 }
 
-module.exports.run = function()
-{
-	successes = 0
-	failures = 0
-	console.log('Test: belongsto')
-	condition(1108730699, [102191581, 85633287, 85688877])
-	console.log('Successes:'.green, successes, 'Failures:'.red, failures)
-}
+//module.exports.run = function()
+//{
+//	successes = 0
+//	failures = 0
+//	console.log('Test: belongsto')
+//	condition(1108730699, [102191581, 85633287, 85688877])
+//	console.log('Successes:'.green, successes, 'Failures:'.red, failures)
+//}
